+++ app/engine/shaders/model_entity_shadow_shader.vert
@@ -13,7 +13,8 @@ uniform float u_minZ;
 uniform float u_maxX;
 uniform float u_maxY;
 uniform float u_maxZ;
-uniform float u_textureRepeat;
+
+uniform int u_textureRepeat;
 
 out vec2 f_uv;
 
@@ -22,7 +23,7 @@ void main()
 	vec4 worldSpacePosition = (u_transformation * vec4(v_position, 1.0f));
 	vec4 shadowSpacePosition = (u_shadowProjection * u_shadowView * worldSpacePosition);
 
-	f_uv = (v_uv * u_textureRepeat);
+	f_uv = (v_uv * float(u_textureRepeat));
 
 	gl_Position = shadowSpacePosition;
 	gl_ClipDistance[0] = dot(worldSpacePosition, vec4( 1.0f,  0.0f,  0.0f, -u_minX));
