+++ app/engine/shaders/quad3d_entity_shadow_shader.vert
@@ -16,7 +16,8 @@ uniform float u_minZ;
 uniform float u_maxX;
 uniform float u_maxY;
 uniform float u_maxZ;
-uniform float u_textureRepeat;
+
+uniform int u_textureRepeat;
 
 out vec2 f_uv;
 
@@ -25,8 +26,8 @@ void main()
 	vec4 worldSpacePosition = (u_transformation * vec4(v_position, 1.0f));
 	vec4 shadowSpacePosition = (u_shadowProjection * u_shadowView * worldSpacePosition);
 
-	f_uv.x = ((u_uvOffset.x + (v_uv.x * u_uvMultiplier.x)) * u_textureRepeat);
-	f_uv.y = ((u_uvOffset.y + (v_uv.y * u_uvMultiplier.y)) * u_textureRepeat);
+	f_uv.x = ((u_uvOffset.x + (v_uv.x * u_uvMultiplier.x)) * float(u_textureRepeat));
+	f_uv.y = ((u_uvOffset.y + (v_uv.y * u_uvMultiplier.y)) * float(u_textureRepeat));
 
 	gl_Position = shadowSpacePosition;
 	gl_ClipDistance[0] = dot(worldSpacePosition, vec4( 1.0f,  0.0f,  0.0f, -u_minX));
