+++ src/script_interpreter_func_fe3d_quad2d_get.cpp
@@ -205,9 +205,9 @@ const bool ScriptInterpreter::_executeFe3dQuad2dGetter(const string& functionNam
 		{
 			if(_validateFe3dQuad2d(args[0]->getString(), false))
 			{
-				const auto result = _fe3d->quad2d_getTextureRepeat(args[0]->getString());
+				const auto result = static_cast<int>(_fe3d->quad2d_getTextureRepeat(args[0]->getString()));
 
-				returnValues.push_back(make_shared<ScriptValue>(SVT::DECIMAL, result));
+				returnValues.push_back(make_shared<ScriptValue>(SVT::INTEGER, result));
 			}
 		}
 	}
